requirements.txt
aiohttp==3.8.5
aiosignal==1.3.1
asgiref==3.5.2
async-timeout==4.0.2
attrs==23.1.0
autopep8==1.7.0
certifi==2022.12.7
charset-normalizer==3.2.0
colorama==0.4.6
dataclasses-json==0.5.14
distlib==0.3.6
dj-database-url==1.2.0
Django==4.1
django-cors-headers==3.13.0
django-debug-toolbar==3.6.0
django-heroku==0.3.1
djangorestframework==3.13.1
filelock==3.9.0
frozenlist==1.4.0
greenlet==2.0.2
gunicorn==20.1.0
idna==3.4
langchain==0.0.257
langsmith==0.0.19
marshmallow==3.20.1
multidict==6.0.4
mypy-extensions==1.0.0
numexpr==2.8.5
numpy==1.25.2
openai==0.27.8
openapi-schema-pydantic==1.2.4
packaging==23.1
Pillow==9.3.0
pipenv==2022.12.19
platformdirs==2.6.2
psycopg2==2.9.5
psycopg2-binary==2.9.6
pycodestyle==2.9.1
pydantic==1.10.12
python-decouple==3.6
pytz==2022.2.1
PyYAML==6.0.1
requests==2.31.0
SQLAlchemy==2.0.19
sqlparse==0.4.2
tenacity==8.2.2
toml==0.10.2
tqdm==4.65.1
typing-inspect==0.9.0
typing_extensions==4.7.1
tzdata==2022.2
urllib3==2.0.4
virtualenv==20.17.1
virtualenv-clone==0.5.7
whitenoise==6.2.0
yarl==1.9.2
